gsc-bot/requirements.txt
langchain==0.1.20
langchain-openai==0.0.8
langchain-community==0.0.38
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
chromadb==0.4.22
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.1
sqlalchemy==2.0.25
alembic==1.13.1
pandas==2.1.4
numpy==1.26.2
loguru==0.7.2
aiohttp==3.9.1
asyncio==3.4.3
tenacity==8.2.3